+++ spec/dummy_app/app/filters/application_filter.rb
@@ -3,5 +3,4 @@
 require_relative "../../../../lib/tiny_filter/base"
 
 class ApplicationFilter < TinyFilter::Base
-
 end