+++ spec/dummy_app/app/models/post.rb
@@ -4,11 +4,8 @@ require_relative "application_record"
 
 class Post < ApplicationRecord
   class << self
-
     def filter_class
       CustomPostFilter
     end
-
   end
-
 end